feat: refactor waffle
